fix: loss
